+++ private/language/english.php
@@ -2954,6 +2954,7 @@ $LANG_confignames['Core'] = array(
     'cache_memcached_info' => 'Memcached Driver Attributes',
     'default_story_editor' => 'Default Story Editor',
     'story_geshi_enabled' => 'Geshi Code Formatting',
+    'syntax_highlight' => 'Enable Syntax Highlighting / Code Formatting',
 
 // embedded field sets
     'fset_story_display' => 'Story Display',